split UsersClient in two class
